MigrationOptions.ToVersion support long version numbers.

+++ Raven.Migrations/MigrationOptions.cs
@@ -21,7 +21,7 @@ namespace Raven.Migrations
         public IList<Assembly> Assemblies { get; set; }
         public IList<string> Profiles { get; set; }
         public IMigrationResolver MigrationResolver { get; set; }
-        public int ToVersion { get; set; }
+        public long ToVersion { get; set; }
         public MigrationConventions Conventions { get; set; }
         public IMigrationRecordStore MigrationRecordStore { get; set; }
     }
